feat: modularize theme editor
