+++ frontends/nextjs/src/components/editors/theme/constants.ts
@@ -1,0 +1,41 @@
+import { ThemeColors } from './types'
+
+export const DEFAULT_LIGHT_THEME: ThemeColors = {
+  background: 'oklch(0.92 0.03 290)',
+  foreground: 'oklch(0.25 0.02 260)',
+  card: 'oklch(1 0 0)',
+  cardForeground: 'oklch(0.25 0.02 260)',
+  primary: 'oklch(0.55 0.18 290)',
+  primaryForeground: 'oklch(0.98 0 0)',
+  secondary: 'oklch(0.35 0.02 260)',
+  secondaryForeground: 'oklch(0.90 0.01 260)',
+  muted: 'oklch(0.95 0.02 290)',
+  mutedForeground: 'oklch(0.50 0.02 260)',
+  accent: 'oklch(0.70 0.17 195)',
+  accentForeground: 'oklch(0.2 0.02 260)',
+  destructive: 'oklch(0.55 0.22 25)',
+  destructiveForeground: 'oklch(0.98 0 0)',
+  border: 'oklch(0.85 0.02 290)',
+  input: 'oklch(0.85 0.02 290)',
+  ring: 'oklch(0.70 0.17 195)',
+}
+
+export const DEFAULT_DARK_THEME: ThemeColors = {
+  background: 'oklch(0.145 0 0)',
+  foreground: 'oklch(0.985 0 0)',
+  card: 'oklch(0.205 0 0)',
+  cardForeground: 'oklch(0.985 0 0)',
+  primary: 'oklch(0.922 0 0)',
+  primaryForeground: 'oklch(0.205 0 0)',
+  secondary: 'oklch(0.269 0 0)',
+  secondaryForeground: 'oklch(0.985 0 0)',
+  muted: 'oklch(0.269 0 0)',
+  mutedForeground: 'oklch(0.708 0 0)',
+  accent: 'oklch(0.269 0 0)',
+  accentForeground: 'oklch(0.985 0 0)',
+  destructive: 'oklch(0.704 0.191 22.216)',
+  destructiveForeground: 'oklch(0.98 0 0)',
+  border: 'oklch(1 0 0 / 10%)',
+  input: 'oklch(1 0 0 / 15%)',
+  ring: 'oklch(0.556 0 0)',
+}
