+++ frontends/nextjs/src/components/editors/theme/types.ts
@@ -1,0 +1,25 @@
+export interface ThemeColors {
+  background: string
+  foreground: string
+  card: string
+  cardForeground: string
+  primary: string
+  primaryForeground: string
+  secondary: string
+  secondaryForeground: string
+  muted: string
+  mutedForeground: string
+  accent: string
+  accentForeground: string
+  destructive: string
+  destructiveForeground: string
+  border: string
+  input: string
+  ring: string
+}
+
+export interface ThemeConfig {
+  light: ThemeColors
+  dark: ThemeColors
+  radius: string
+}
